requirements.txt
Django==2.0.6
django-ckeditor==5.4.0
django-js-asset==1.1.0
html5lib==1.0.1
httplib2==0.11.3
pdfkit==0.6.1
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyPDF2==1.26.0
pytz==2018.4
reportlab==3.4.0
six==1.11.0
webencodings==0.5.1
xhtml2pdf==0.2.2
